+++ lib/src/storage/dialog_store.dart
@@ -1,4 +1,7 @@
-abstract class ChatterboxStore {
+/// Storage to keep information on pending messages
+abstract class PendingMessagesStore {
+  const PendingMessagesStore();
+
   Future<String?> retrievePending(int userId);
 
   Future<void> setPending(int userId, String stepUrl);